+++ mobile/lib/features/bana_ozel/domain/repositories/bana_ozel_repository.dart
@@ -1,0 +1,9 @@
+import '../entities/bana_ozel_entities.dart';
+
+abstract class BanaOzelRepository {
+  Future<BanaOzelCatalogEntity> fetchCatalog();
+
+  Future<BanaOzelOpenResultEntity> openItem({
+    required BanaOzelItemEntity item,
+  });
+}